move check_metrics to check file
