+++ specparam/metrics/__init__.py
@@ -1,3 +1,4 @@
 """Metrics sub-module."""
 
 from .metric import Metric
+from .check import check_metrics
