+++ specparam/metrics/definitions.py
@@ -75,13 +75,4 @@ METRICS = {
 
 }
 
-
-def check_metrics():
-    """Check the set of available metrics."""
-
-    print('Available metrics:')
-    for metric in METRICS.values():
-        print('    {:8s} {:12s} : {:s}'.format(metric.category, metric.measure, metric.description))
-
-
 check_metric_definition = partial(check_selection, options=METRICS, definition=Metric)
