+++ specparam/tests/metrics/test_check.py
@@ -1,0 +1,10 @@
+"""Test functions for specparam.metrics.check."""
+
+from specparam.metrics.check import *
+
+###################################################################################################
+###################################################################################################
+
+def test_check_metrics():
+
+    check_metrics()
