+++ specparam/tests/metrics/test_definitions.py
@@ -12,7 +12,3 @@ def test_metrics_library():
     for key, metric in METRICS.items():
         assert isinstance(metric, Metric)
         assert metric.label == key
-
-def test_check_metrics():
-
-    check_metrics()
